+++ src/System.Security.Permissions/ref/System.Security.Permissions.Forwards.cs
@@ -18,3 +18,4 @@
 [assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Security.PermissionSet))]
 [assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.Security.Permissions.PermissionState))]
 #endif
+
