+++ src/System.Security.Permissions/tests/PermissionTests.cs
@@ -7,7 +7,6 @@ using System.Configuration;
 using System.DirectoryServices;
 using System.Reflection;
 using System.Web;
-using System.Xaml.Permissions;
 
 namespace System.Security.Permissions.Tests
 {
@@ -442,26 +441,6 @@ namespace System.Security.Permissions.Tests
         {
             WebBrowserPermissionAttribute wpa = new WebBrowserPermissionAttribute(new Permissions.SecurityAction());
             IPermission ip = wpa.CreatePermission();
-        }
-
-        [Fact]
-        public static void XamlLoadPermissionCallMethods()
-        {
-            XamlAccessLevel accessLevel = XamlAccessLevel.AssemblyAccessTo(Assembly.GetExecutingAssembly().GetName());
-            XamlLoadPermission xp = new XamlLoadPermission(accessLevel);
-            XamlLoadPermission xp2 = new XamlLoadPermission(PermissionState.Unrestricted);
-            XamlLoadPermission xp3 = new XamlLoadPermission(Array.Empty<XamlAccessLevel>());
-            bool testbool = xp.IsUnrestricted();
-            IPermission ip = xp.Copy();
-            IPermission ip2 = xp.Intersect(ip);
-            IPermission ip3 = xp.Union(ip);
-            testbool = xp.IsSubsetOf(ip);
-            testbool = xp.Equals(new object());
-            testbool = xp.Includes(accessLevel);
-            int hash = xp.GetHashCode();
-            SecurityElement se = new SecurityElement("");
-            xp.FromXml(se);
-            se = xp.ToXml();
         }
 
         [Fact]
